Data type change in service code

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double a, double b) {
+    public int calculate(double a, double b) {
         double result =  (a / (b * b));
         return (int) result;
     }
